requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.26.0
requests>=2.32.0
PyMuPDF>=1.24.0
beautifulsoup4>=4.12.0
trafilatura>=1.9.0
